+++ src/platform_impl/macos/view.rs
@@ -239,7 +239,10 @@ declare_class!(
             // 2. Even when a window resize does occur on a new tabbed window, it contains the wrong size (includes tab height).
             let logical_size = LogicalSize::new(rect.size.width as f64, rect.size.height as f64);
             let size = logical_size.to_physical::<u32>(self.scale_factor());
-            self.queue_event(WindowEvent::Resized(size));
+            self.queue_event(WindowEvent::Configured {
+                size,
+                state: todo!(),
+            });
         }
 
         #[sel(drawRect:)]