+++ src/platform_impl/web/event_loop/window_target.rs
@@ -713,7 +713,11 @@ impl<T> EventLoopWindowTarget<T> {
                         canvas.set_old_size(new_size);
                         runner.send_event(Event::WindowEvent {
                             window_id: RootWindowId(id),
-                            event: WindowEvent::Resized(new_size),
+                            event: WindowEvent::Configured {
+                                size: new_size,
+                                // Fullscreen?
+                                state: todo!(),
+                            },
                         });
                         runner.request_redraw(RootWindowId(id));
                     }